Add set_options operations
